Refactoring WIP (dedicated services, pages separated from components)
1/ AdminService & Admin pages (GameAccount/Character, Crew)
