+++ src/SWGIndustries/Components/GameAccounts/AddNewGameAccount.razor
@@ -1,14 +1,14 @@
 ﻿@using SWGIndustries.Data
 @using SWGIndustries.Services
 
-@inject DataAccessService DataAccessService
+@inject AdminService AdminService
 @inject ISnackbar Snackbar
 
 <MudPaper Class="pa-4" Style="width: 50%">
     <MudText Typo="Typo.h6">Add a new Game Account</MudText>
     <MudForm @ref="_accountForm" @bind-IsValid="@_formSuccess" Model="@_newAccount">
         <MudTextField T="string" Label="Name" Required="true" AutoFocus="true" Immediate="true" DebounceInterval="100"
-                      Validation="@(new Func<string, string>(ValidateAccountName))" @bind-Value="@_newAccount.Name"></MudTextField>
+                      Validation="@(async (string t) => await ValidateAccountName(t))" @bind-Value="@_newAccount.Name"></MudTextField>
         <div class="d-flex justify-end mt-4">
             <MudButton Variant="Variant.Filled" Color="Color.Secondary" Class="mr-2"
                        OnClick="@(() => OnAccountAdded?.Invoke(null))">Cancel</MudButton>
@@ -29,7 +29,7 @@
     
     private async Task CreateGameAccount()
     {
-        if (await DataAccessService.AddGameAccount(_newAccount, true))
+        if (await AdminService.AddGameAccount(_newAccount, true))
         {
             Snackbar.Add($"Game Account '{_newAccount.Name}' added successfully.", Severity.Success);
         }
@@ -42,14 +42,14 @@
         _newAccount = new GameAccountEntity();
     }
 
-    private string ValidateAccountName(string name)
+    private async Task<string> ValidateAccountName(string name)
     {
         if (string.IsNullOrWhiteSpace(name))
         {
             return "Name is required.";
         }
 
-        if (DataAccessService.DbContext.GameAccounts.Any(c => c.Name == name))
+        if (await AdminService.GetGameAccountByName(name, true) != null)
         {
             return "Name already exists.";
         }
